pyproject.toml
[project]
name = "support-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
